requirements.txt
black>=22.1.0
flake8>=4.0.1
hydra-core>=1.1.1
mypy>=0.931
pre-commit>=2.17.0
pytest>=6.2.5
pytorch-lightning~=1.5.9
torchvision>=0.8.2
